Post_CRUD/requirements.txt
Django==2.0.5
django-jet==1.0.8
django-widget-tweaks==1.4.3
djangorestframework==3.11.1
pytz==2018.5
